# cmake: try well-known send/recv signature for Apple